+++ app/views/Basket.html
@@ -33,7 +33,7 @@
             </table>
 
             <div class="container-fluid well">
-                <button type="submit" class="btn btn-default" ng-disabled="products.length < 1" ng-click="order()">Submit Order</button>
+                <button type="submit" class="btn btn-default" ng-disabled="products.length < 1" ng-click="checkout()">Checkout</button>
             </div>
         </div>
     </div>
